Made the changes suggested by reviewers in test folder:
