+++ test/leader-board-call.test.js
@@ -1,4 +1,5 @@
-import { submitHighScore, getScoreBoard, createGame } from './leader-board-call-mock';
+import {createGame} from './create-game-mock';
+import { submitScore, getScoreBoard } from '../src/leader-board-call';
 
 describe('Test to add a game, add a score to it and request it back', () => {
   test('Add a mock game and receive a message with the id', () => {
@@ -11,7 +12,7 @@ describe('Test to add a game, add a score to it and request it back', () => {
       test('Add a record for the previous id', () => {
         const user = 'UserName';
         const score = 5000;
-        const result2 = submitHighScore(user, score, id);
+        const result2 = submitScore(user, score, id);
         result2.then(answer2 => {
           expect(answer2).toBe('Leaderboard score created correctly.');
         });
