--- test/mocks/styleMock.js
@@ -1,1 +1,0 @@
-module.exports = {};
